backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
psycopg[binary]==3.1.18
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.1.0
gunicorn==21.2.0
Authlib==1.3.0
requests==2.31.0

# CV Text Extraction dependencies
python-docx==1.1.0
pdfminer.six==20221105
pdf2image==1.16.3
pytesseract==0.3.10
Pillow>=10.4.0
reportlab==4.0.7

# Auto Scraper dependencies
selenium==4.16.0
beautifulsoup4==4.12.2
pandas>=2.2.0
APScheduler==3.10.4
openpyxl==3.1.2
